Adding JSON Path Parameters, Conditional Expressions and Converters
These features allow a bit more logic when deciding to select source values and provides the ability to return default values embedded as part of the path.  Converters allows functions to be chained together to alter the selected value. There is also a mechanism for adding custom Converters similar to extending Transforms.